Remove root package and move node_modules to relevant directories
