+++ client/.env
@@ -1,3 +1,11 @@
 SKIP_PREFLIGHT_CHECK=true
 REACT_APP_SERVER_HOST=localhost
 REACT_APP_SERVER_PORT=3001
+DB_URL=mongodb://127.0.0.1
+DB_PORT=27017
+DB_NAME=leitner
+SERVER_HOST=localhost
+SERVER_PORT=3001
+CLIENT_HOST=localhost
+CLIENT_PORT=3000
+TEST_ENV=true
