+++ client/src/components/AddQuiz/AddQuiz.test.tsx
@@ -4,7 +4,7 @@ import '@testing-library/jest-dom/extend-expect';
 import userEvent from '@testing-library/user-event';
 import AddQuiz from './AddQuiz';
 import { addCard, getAllCards } from '../../utils/axios';
-import { dropAllTestCollections } from '../../../../testUtils/testUtils';
+import { dropAllTestCollections } from '../../../testUtils/testUtils';
 
 beforeAll(async () => { await dropAllTestCollections()});
 afterEach(async () => { await dropAllTestCollections()});
