+++ client/src/components/EditQuiz/EditQuiz.test.tsx
@@ -4,7 +4,7 @@ import userEvent from '@testing-library/user-event'
 import '@testing-library/jest-dom/extend-expect';
 import EditQuiz from './EditQuiz';
 import { addCard, getAllCards, addQuiz, getAllQuizzes } from '../../utils/axios';
-import { dropAllTestCollections } from '../../../../testUtils/testUtils';
+import { dropAllTestCollections } from '../../../testUtils/testUtils';
 
 beforeAll(async () => { await dropAllTestCollections() });
 afterEach(async () => { await dropAllTestCollections() });
