+++ client/src/screens/Study/Study.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { Quiz, FormattedCard, QuizUnderstudy } from '../../../../types';
+import { Quiz, FormattedCard, QuizUnderstudy } from '../../../types';
 import StudyHome from '../../components/StudyHome/StudyHome';
 import StudyQuestion from '../../components/StudyQuestion/StudyQuestion';
 import StudyReview from '../../components/StudyReview/StudyReview';
