--- package.json
@@ -1,59 +1,0 @@
-{
-  "name": "leitner",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "npx jest -i src/",
-    "test:e2e:firefox": "npx testcafe firefox e2e/**/*.ts",
-    "test:e2e:chrome": "npx testcafe chrome e2e/**/*.ts",
-    "tsc": "npx tsc",
-    "lint": "npx eslint . --ext .js,.ts,.tsx --fix",
-    "start:server": "npx ts-node server/index.ts",
-    "start:client": "cd client && npm run start",
-    "prepare": "husky install"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "1.2.35",
-    "@fortawesome/free-solid-svg-icons": "5.15.3",
-    "@fortawesome/react-fontawesome": "0.1.14",
-    "axios": "0.21.1",
-    "dotenv": "8.2.0",
-    "express": "4.17.1",
-    "mongodb": "3.6.4"
-  },
-  "devDependencies": {
-    "@babel/cli": "7.14.8",
-    "@babel/core": "7.15.0",
-    "@babel/preset-env": "7.15.0",
-    "@babel/preset-typescript": "7.15.0",
-    "@testing-library/jest-dom": "5.14.1",
-    "@testing-library/react": "12.0.0",
-    "@types/express": "4.17.13",
-    "@types/jest": "27.0.1",
-    "@types/mongodb": "3.6.10",
-    "@types/node": "16.6.1",
-    "@types/react": "17.0.18",
-    "@types/react-dom": "17.0.9",
-    "@typescript-eslint/eslint-plugin": "^5.29.0",
-    "@typescript-eslint/parser": "^5.29.0",
-    "babel-jest": "27.0.6",
-    "eslint": "^8.18.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jest": "24.4.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.30.0",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-testcafe": "0.2.1",
-    "husky": "^7.0.1",
-    "jest": "27.0.6",
-    "nodemon": "2.0.12",
-    "testcafe": "1.15.2",
-    "ts-jest": "^27.1.5",
-    "ts-node": "^10.2.0",
-    "typescript": "^4.3.5"
-  }
-}
